requirements.txt
appdirs==1.4.0
astroid==1.4.9
Django==1.10.5
django-filter==1.0.1
djangorestframework==3.5.4
isort==4.2.5
lazy-object-proxy==1.2.2
Markdown==2.6.8
mccabe==0.6.1
packaging==16.8
psycopg2==2.6.2
pylint==1.6.5
pyparsing==2.1.10
six==1.10.0
wrapt==1.10.8
